refactor(tests): centralize synth_docx fixture in docspec-test-utils
