+++ crates/docspec-cli/Cargo.toml
@@ -31,7 +31,7 @@ predicates = "3"
 tempfile = "3"
 serde_json = "1"
 reqwest = { version = "0.13", default-features = false, features = ["blocking"] }
-zip = { version = "8", default-features = false, features = ["deflate"] }
+docspec-test-utils = { workspace = true }
 
 [build-dependencies]
 clap = { version = "4", features = ["derive"] }
