+++ crates/docspec-http/Cargo.toml
@@ -54,7 +54,7 @@ tower = { version = "0.5", features = ["util"] }
 serde_json = "1"
 reqwest = { version = "0.13", features = ["blocking", "json"] }
 uuid = "1"
-zip = { version = "8", default-features = false, features = ["deflate"] }
+docspec-test-utils = { workspace = true }
 sentry = { version = "0.48", default-features = false, features = [
   "backtrace",
   "contexts",
