+++ crates/docspec-test-utils/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "docspec-test-utils"
+version.workspace = true
+publish = false
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "Internal test fixtures shared across the docspec workspace. Not published."
+
+[dependencies]
+zip = { version = "8", default-features = false, features = ["deflate"] }
+
+[lints]
+workspace = true
