+++ crates/docspec/Cargo.toml
@@ -52,7 +52,7 @@ docspec-pandoc-native-writer = { workspace = true, optional = true }
 [dev-dependencies]
 serde_json = "1"
 tempfile = "3"
-zip = { version = "8", default-features = false, features = ["deflate"] }
+docspec-test-utils = { workspace = true }
 
 [package.metadata.docs.rs]
 all-features = true
