+++ crates/docspec/tests/docx_color_to_blocknote.rs
@@ -18,28 +18,13 @@
     clippy::doc_paragraphs_missing_punctuation
 )]
 
-use std::io::{Cursor, Write as _};
-
-use zip::{write::SimpleFileOptions, CompressionMethod, ZipWriter};
+use std::io::Cursor;
 
 use docspec::writers::BlockNoteWriter;
 use docspec::{AnyReader, EventSink as _, EventSource as _, InputFormat, StackTrackingSink};
+use docspec_test_utils::synth_docx;
 
 const SIMPLE_RELS: &str = r#"<?xml version="1.0"?><Relationships xmlns="http://schemas.openxmlformats.org/package/2006/relationships"><Relationship Id="rId1" Type="http://schemas.openxmlformats.org/officeDocument/2006/relationships/officeDocument" Target="word/document.xml"/></Relationships>"#;
-
-fn synth_docx(rels_xml: &str, document_xml: &str) -> Vec<u8> {
-    let buf = Cursor::new(Vec::new());
-    let mut zip_writer = ZipWriter::new(buf);
-    let opts = SimpleFileOptions::default().compression_method(CompressionMethod::Deflated);
-    zip_writer.start_file("_rels/.rels", opts).unwrap();
-    zip_writer.write_all(rels_xml.as_bytes()).unwrap();
-    let opts_doc = SimpleFileOptions::default().compression_method(CompressionMethod::Deflated);
-    zip_writer
-        .start_file("word/document.xml", opts_doc)
-        .unwrap();
-    zip_writer.write_all(document_xml.as_bytes()).unwrap();
-    zip_writer.finish().unwrap().into_inner()
-}
 
 /// End-to-end test: DOCX with color, highlight, and shading → BlockNote JSON.
 ///
