+++ crates/docspec/tests/readers_docx.rs
@@ -41,22 +41,11 @@ mod tests {
     }
 }
 
-use std::io::{Cursor, Write as _};
-use zip::{write::SimpleFileOptions, CompressionMethod, ZipWriter};
+use std::io::Cursor;
+
+use docspec_test_utils::synth_docx;
 
 const SIMPLE_RELS: &str = r#"<?xml version="1.0"?><Relationships xmlns="http://schemas.openxmlformats.org/package/2006/relationships"><Relationship Id="rId1" Type="http://schemas.openxmlformats.org/officeDocument/2006/relationships/officeDocument" Target="word/document.xml"/></Relationships>"#;
-
-fn synth_docx(rels_xml: &str, document_xml: &str) -> Vec<u8> {
-    let buf = Cursor::new(Vec::new());
-    let mut writer = ZipWriter::new(buf);
-    let opts = SimpleFileOptions::default().compression_method(CompressionMethod::Deflated);
-    writer.start_file("_rels/.rels", opts).unwrap();
-    writer.write_all(rels_xml.as_bytes()).unwrap();
-    let opts_doc = SimpleFileOptions::default().compression_method(CompressionMethod::Deflated);
-    writer.start_file("word/document.xml", opts_doc).unwrap();
-    writer.write_all(document_xml.as_bytes()).unwrap();
-    writer.finish().unwrap().into_inner()
-}
 
 #[test]
 fn any_reader_from_reader_docx_hello() {
